0.3.0

+++ src/ConePackage.class.php
@@ -47,6 +47,37 @@ class ConePackage
 		{
 			switch($step["type"])
 			{
+				case "platform_switch":
+					if(Cone::isWindows())
+					{
+						if(array_key_exists("windows", $step))
+						{
+							$this->performSteps($step["windows"]);
+						}
+					}
+					else
+					{
+						if(array_key_exists("unix", $step))
+						{
+							$this->performSteps($step["unix"]);
+						}
+						if(Cone::isLinux())
+						{
+							if(array_key_exists("linux", $step))
+							{
+								$this->performSteps($step["linux"]);
+							}
+						}
+						else if(Cone::isMacOS())
+						{
+							if(array_key_exists("macos", $step))
+							{
+								$this->performSteps($step["macos"]);
+							}
+						}
+					}
+					break;
+
 				case "shell_exec":
 					shell_exec($step["value"]);
 					break;
@@ -64,41 +95,60 @@ class ConePackage
 					break;
 
 				case "download":
-					file_put_contents($step["name"], file_get_contents($step["url"]));
+					Cone::download($step["url"], $step["name"]);
+					if(array_key_exists("hash", $step))
+					{
+						foreach($step["hash"] as $algo => $hash)
+						{
+							if(hash_file($algo, $step["name"]) != $hash)
+							{
+								echo $step["name"]." signature mismatch, aborting install.\n";
+								unlink($step["name"]);
+								return;
+							}
+						}
+					}
+					break;
+
+				case "extract":
+					if(!is_file($step["file"]))
+					{
+						echo "Warning: ".$step["file"]." can't be extracted as it doesn't exist.\n";
+						break;
+					}
+					Cone::extract($step["file"], $step["target"]);
 					break;
 
 				case "delete":
-					if(!file_exists($step["value"]))
+					if(!file_exists($step["file"]))
 					{
-						echo "Warning: ".$step["value"]." can't be deleted as it doesn't exist.\n";
+						echo "Warning: ".$step["file"]." can't be deleted as it doesn't exist.\n";
 						break;
 					}
-					Cone::reallyDelete($step["value"]);
+					Cone::reallyDelete($step["file"]);
 					break;
 
 				case "keep":
-					if(!file_exists($step["value"]))
+					if(!file_exists($step["file"]))
 					{
-						echo "Warning: ".$step["value"]." can't be kept as it doesn't exist.\n";
+						echo "Warning: ".$step["file"]." can't be kept as it doesn't exist.\n";
 						break;
 					}
-					$dir = __DIR__."/../packages/";
-					if(!is_dir($dir))
+					$dir = __DIR__."/../packages/".$this->name."/";
+					if(!is_dir($dir) && $step["as"] != "")
 					{
 						mkdir($dir);
 					}
-					$dir = $dir.$this->name."/";
-					if(!is_dir($dir))
-					{
-						mkdir($dir);
-					}
-					rename($step["value"], $dir.$step["value"]);
+					rename($step["file"], $dir.$step["as"]);
 					break;
+
+				default:
+					echo "Error: Unknown step type: ".$step["type"]."\n";
 			}
 		}
 	}
 
-	function install(&$installed_packages, $indents = 0)
+	function install(&$installed_packages, $dependency_of = null)
 	{
 		if($this->isInstalled())
 		{
@@ -117,28 +167,39 @@ class ConePackage
 							return;
 						}
 						break;
+
+					default:
+						echo "Error: Unknown prerequisite type: ".$prerequisite["type"]."\n";
 				}
 			}
 		}
-		if($indents > 0)
-		{
-			echo str_repeat("\t", $indents)."- Installing dependency ".$this->name."...\n";
-		}
-		else
+		if($dependency_of === null)
 		{
 			echo "Installing ".$this->name."...\n";
 		}
-		foreach($this->getDependencies() as $dependency)
+		else
 		{
-			$dependency->install($installed_packages, $indents + 1);
+			echo "Installing ".$dependency_of." dependency ".$this->name."...\n";
+		}
+		if(array_key_exists("dependencies", $this->data))
+		{
+			foreach($this->getDependencies() as $dependency)
+			{
+				$dependency->install($installed_packages, $this->name);
+			}
+			echo "All ".$this->name." dependencies are installed.\n";
+		}
+		if(!is_dir(__DIR__."/../packages/"))
+		{
+			mkdir(__DIR__."/../packages/");
 		}
 		if(array_key_exists("install", $this->data))
 		{
 			$this->performSteps($this->data["install"]);
 		}
+		$working_directory = realpath(__DIR__."/../packages/".$this->name);
 		if(array_key_exists("shortcuts", $this->data))
 		{
-			$working_directory = realpath(__DIR__."/../packages/".$this->name);
 			if($working_directory === false)
 			{
 				echo "Warning: Can't create any shortcuts as no file was kept.\n";
@@ -147,11 +208,44 @@ class ConePackage
 			{
 				foreach($this->data["shortcuts"] as $shortcut)
 				{
-					Cone::createPathShortcut($shortcut["name"], Cone::which($shortcut["target_which"]), $shortcut["target_arguments"], realpath(__DIR__."/../packages/".$this->name));
+					$options = [
+						"working_directory" => $working_directory
+					];
+					if(array_key_exists("target_arguments", $shortcut))
+					{
+						$options["target_arguments"] = $shortcut["target_arguments"];
+					}
+					if(array_key_exists("target", $shortcut))
+					{
+						$options["target"] = realpath($working_directory."/".$shortcut["target"].(Cone::isWindows() ? ".exe" : ""));
+					}
+					else if(array_key_exists("target_which", $shortcut))
+					{
+						$options["target"] = Cone::which($shortcut["target_which"]);
+					}
+					else
+					{
+						echo "Error: Shortcut is missing target or target_which.\n";
+					}
+					Cone::createPathShortcut($shortcut["name"], $options);
 				}
 			}
 		}
-		$installed_packages[$this->name] = ["manual" => $indents == 0];
+		if(Cone::isWindows() && array_key_exists("file_associations", $this->data))
+		{
+			if($working_directory === false)
+			{
+				echo "Warning: Can't create any file associations as no file was kept.\n";
+			}
+			else
+			{
+				foreach($this->data["file_associations"] as $ext => $cmd)
+				{
+					shell_exec("Assoc .{$ext}={$ext}file\nFtype {$ext}file={$working_directory}\\{$cmd}");
+				}
+			}
+		}
+		$installed_packages[$this->name] = ["manual" => ($dependency_of === null)];
 		if(array_key_exists("version", $this->data))
 		{
 			$installed_packages[$this->name]["version"] = $this->data["version"];
@@ -173,9 +267,19 @@ class ConePackage
 		{
 			Cone::reallyDelete($dir);
 		}
-		foreach($this->data["shortcuts"] as $shortcut)
+		if(array_key_exists("shortcuts", $this->data))
 		{
-			unlink(Cone::getPathFolder().$shortcut["name"].(Cone::isWindows() ? ".lnk" : ""));
+			foreach($this->data["shortcuts"] as $shortcut)
+			{
+				unlink(Cone::getPathFolder().$shortcut["name"].(Cone::isWindows() ? ".lnk" : ""));
+			}
+		}
+		if(Cone::isWindows() && array_key_exists("file_associations", $this->data))
+		{
+			foreach($this->data["file_associations"] as $ext => $cmd)
+			{
+				shell_exec("Ftype {$ext}file=\nAssoc .{$ext}=");
+			}
 		}
 		if(array_key_exists("uninstall", $this->data))
 		{
